+++ aws-py-apigateway-lambda-serverless/hello_lambda/hello.py
@@ -1,0 +1,6 @@
+import json
+
+def handler(event, context):
+    return {
+          "statusCode": 200,
+          "body": "Hello World from Pulumi in python via AWS Lambda!!!"}